refactor: migrate SteamPunkGame to Kotlin

+++ src/main/java/gruvexp/bbminigames/twtClassic/botbowsGames/BotBowsGame.kt
@@ -294,7 +294,7 @@ open class BotBowsGame(val settings: Settings) {
 
         val statsLocY = statsLocation.blockY
         statsLocation.add(statsLocation.getDirection().multiply(10))
-        statsLocation.y = (statsLocY + 3).toDouble()
+        statsLocation.y = statsLocY + 3.0
 
         val resultDisplay = ResultDisplay(statsLocation, matchResult)
         Bukkit.getScheduler().runTaskLater(Main.getPlugin(), Runnable { resultDisplay.remove() }, 60 * 20L)
